# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'ModWorkshop API Documentation'
  description: "The ModWorkshop API is available for everyone to use. \r\n\r\nDo note that by using the API you must follow the following guidelines:\r\n\r\n1. Do not spam the API.\r\n\r\n2. Do not replicate the site or remove the need to visit the site to download mods. \r\nExceptions: You are allowed to implement updates, you are allowed to make a mod downloadable to ease the process of joining a game with mods (Example: downloading maps). Integration to the site itself is fine too (We'll add ways to download mods directly to mod managers soon).\r\n\r\n3. Respect user's privacy and do not store their information without their consent.\r\n\r\nAt the moment, the API <b>only supports GET requests</b>. More support will come in the future, but will require the use of API keys."
  version: 1.0.0
servers:
  -
    url: 'https://api.modworkshop.net'
paths:
...
  '/games/{game_id}/users/{user_id}/roles':
    patch:
      summary: 'Set user roles'
      operationId: setUserRoles
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Games
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                role_ids:
                  type: array
                  description: 'Must be at least 1.'
                  example:
                    - 57
                  items:
                    type: integer
    parameters:
      -
        in: path
        name: game_id
        description: 'The ID of the game.'
        example: 2
        required: true
        schema:
          type: integer
      -
        in: path
        name: user_id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/users/{user}':
    get:
      summary: 'Get User'
      operationId: getUser
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Users
      security: []
    parameters:
      -
        in: path
        name: user
        description: 'The ID of the user'
        example: 15
        required: true
        schema:
          type: integer
  '/users/{user_id}/roles':
    patch:
      summary: 'Set User Roles'
      operationId: setUserRoles
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Users
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                role_ids:
                  type: array
                  description: 'Must be at least 2.'
                  example:
                    - 25
                  items:
                    type: integer
    parameters:
      -
        in: path
        name: user_id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
...